backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
python-multipart==0.0.9
pydantic==2.9.0
supabase==2.9.1
pdfplumber==0.11.0
python-docx==1.1.0
spacy==3.7.5
scikit-learn==1.5.1
numpy==1.26.4
pytest==8.3.2
